refactor: デフォルト辞書データを別ファイルに分離
## 変更内容
- `default-dictionaries.ts`を新規作成
- DEFAULT_TECHNICAL_WORDS、CHARACTER_NAME_WORDSを移動
- ALL_DEFAULT_WORDSを追加（全辞書の統合）
- 関連ファイルのimportを更新

## 影響ファイル
- dictionary-client.ts: 辞書データを削除
- default-dictionaries.ts: 新規作成
- dictionary-register.ts: import更新
- server.ts: import更新
- dictionary-client.test.ts: import更新

## 目的
- 責務の分離（APIクライアントとデータの分離）
- 保守性の向上
- 辞書データの管理を一元化
