+++ src/cli/dictionary-register.ts
@@ -12,13 +12,9 @@
 
 import { Command } from 'commander';
 import fs from 'fs';
-import { 
-  DictionaryClient, 
-  DictionaryWord,
-  DEFAULT_TECHNICAL_WORDS,
-  CHARACTER_NAME_WORDS 
-} from '../core/dictionary/dictionary-client.js';
+import { DictionaryClient, DictionaryWord } from '../core/dictionary/dictionary-client.js';
 import { DictionaryPersistenceManager } from '../core/dictionary/dictionary-persistence.js';
+import { DEFAULT_TECHNICAL_WORDS, CHARACTER_NAME_WORDS } from '../core/dictionary/default-dictionaries.js';
 
 const program = new Command();
